third_party/blink/web_tests/external/wpt/mixed-content/gen/sharedworker-classic.http-rp/unset/fetch.https-expected.txt

This is a testharness.js-based test.
[FAIL] Mixed-Content: Expects blocked for fetch to cross-http origin and keep-scheme redirection from https context.
  assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
[FAIL] Mixed-Content: Expects blocked for fetch to cross-http origin and no-redirect redirection from https context.
  assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
[FAIL] Mixed-Content: Expects blocked for fetch to cross-http origin and swap-scheme redirection from https context.
  assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
[FAIL] Mixed-Content: Expects blocked for fetch to same-http origin and keep-scheme redirection from https context.
  assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
[FAIL] Mixed-Content: Expects blocked for fetch to same-http origin and no-redirect redirection from https context.
  assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
[FAIL] Mixed-Content: Expects blocked for fetch to same-http origin and swap-scheme redirection from https context.
  assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
Harness: the test ran to completion.
